LayoutTests/imported/w3c/web-platform-tests/css/css-contain/container-queries/conditional-container-status-expected.txt

You should see a green border around this text

FAIL Conditionally applying container-type:initial assert_equals: expected "rgb(0, 128, 0)" but got "rgb(0, 0, 0)"
